tests/baseline/protocol-1.2/inspector/console/console-tests-expected.txt

CONSOLE MESSAGE: line 9: log
CONSOLE DEBUG: line 10: debug
CONSOLE INFO: line 11: info
CONSOLE WARNING: line 12: warn
CONSOLE ERROR: line 13: error
CONSOLE MESSAGE: line 15: repeated
CONSOLE MESSAGE: line 15: repeated
CONSOLE MESSAGE: line 15: repeated
CONSOLE MESSAGE: line 15: repeated
CONSOLE MESSAGE: line 15: repeated
CONSOLE DEBUG: line 17: count: 1
CONSOLE DEBUG: line 17: count: 2
CONSOLE MESSAGE: line 18: group
CONSOLE MESSAGE: line 19: console.groupEnd
CONSOLE MESSAGE: line 20: 1
CONSOLE MESSAGE: line 21: groupCollapsed
CONSOLE MESSAGE: line 22: [object Object]
CONSOLE MESSAGE: line 23: 42
CONSOLE MESSAGE: line 24: true
CONSOLE MESSAGE: line 25: null
CONSOLE MESSAGE: line 26: undefined
CONSOLE MESSAGE: line 27: [object HTMLDocument]
CONSOLE MESSAGE: line 28: function () { }
CONSOLE MESSAGE: line 29: function f() { }
CONSOLE MESSAGE: line 30: 1,2,3
CONSOLE MESSAGE: line 31: /regexp.*/
CONSOLE MESSAGE: line 32: console.groupEnd
CONSOLE DEBUG: line 33: : 1
CONSOLE DEBUG: line 34: : 1
CONSOLE DEBUG: line 35: : 1
CONSOLE DEBUG: line 36: title: 1
CONSOLE DEBUG: line 37: title: 2
CONSOLE DEBUG: line 38: title: 3
Tests that console logging dumps proper messages.

console-tests.html:9 log console-message-wrapper console-info-level > console-message
console-tests.html:10 debug console-message-wrapper console-verbose-level > console-message
console-tests.html:11 info console-message-wrapper console-info-level > console-message
console-tests.html:12 warn
onload @ console-tests.html:12 console-message-wrapper console-warning-level > console-message
console-tests.html:13 error
onload @ console-tests.html:13 console-message-wrapper console-error-level > console-message
5console-tests.html:15 repeated console-message-wrapper console-info-level > console-message repeated-message
console-tests.html:17 count: 1 console-message-wrapper console-verbose-level > console-message
console-tests.html:17 count: 2 console-message-wrapper console-verbose-level > console-message
console-tests.html:18 group console-message-wrapper console-info-level > console-message console-group-title
console-tests.html:20 1 2 3 console-message-wrapper console-info-level > console-message
console-tests.html:21 groupCollapsed console-message-wrapper console-info-level > console-message console-group-title
console-tests.html:33 : 1 console-message-wrapper console-verbose-level > console-message
console-tests.html:34 : 1 console-message-wrapper console-verbose-level > console-message
console-tests.html:35 : 1 console-message-wrapper console-verbose-level > console-message
console-tests.html:36 title: 1 console-message-wrapper console-verbose-level > console-message
console-tests.html:37 title: 2 console-message-wrapper console-verbose-level > console-message
console-tests.html:38 title: 3 console-message-wrapper console-verbose-level > console-message
